requirements.txt
anthropic>=0.40.0
streamlit>=1.36.0
google-auth>=2.0
google-auth-oauthlib>=1.0
google-api-python-client>=2.0
reportlab>=4.0
python-dotenv>=1.0
